GAN_word/requirements.txt
alembic==1.16.4
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
bottle==0.13.4
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.2
fsspec==2025.5.1
greenlet==3.2.3
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.34.4
imageio==2.37.0
inquirerpy==0.3.4
Jinja2==3.1.6
joblib==1.5.1
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.1
llvmlite==0.44.0
Mako==1.3.10
matplotlib==3.10.5
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numba==0.61.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.10.0
openpyxl==3.1.5
optuna==4.4.0
optuna-dashboard==0.19.0
pandas==2.3.3
pandocfilters==1.5.1
pfzy==0.3.4
pillow==11.2.1
pipreqs==0.4.13
pyparsing==3.2.3
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
transformers==4.54.0
triton==3.3.1
tzdata==2025.2
webencodings==0.5.1
wheel==0.45.1
yarg==0.1.10
